move FeatureExpectOptionsChooser to logic
