+++ logic/atrium-logic-common/src/main/kotlin/ch/tutteli/atrium/logic/creating/RootExpectBuilder.kt
@@ -146,7 +146,7 @@ interface RootExpectBuilder {
         companion object {
             @ExperimentalNewExpectTypes
             fun <T> createAndBuild(configuration: OptionsChooser<T>.() -> Unit): RootExpectOptions<T> =
-                OptionsChooserImpl<T>().apply(configuration).build()
+                RootExpectOptionsChooserImpl<T>().apply(configuration).build()
         }
     }
 
@@ -189,28 +189,3 @@ interface RootExpectBuilder {
         }
     }
 }
-
-/**
- * Helper function to create a [RootExpectOptions] via
- */
-@ExperimentalNewExpectTypes
-@Suppress("FunctionName")
-fun <T> RootExpectOptions(configuration: RootExpectBuilder.OptionsChooser<T>.() -> Unit): RootExpectOptions<T> =
-    RootExpectBuilder.OptionsChooser.createAndBuild(configuration)
-
-/**
- * Helper function to create a [RootExpectOptions] by specifying components via named parameters.
- *
- * You can use it as follows: `RootExpectOptions(reporter = myReporter)`
- */
-@ExperimentalNewExpectTypes
-@Suppress("FunctionName")
-fun <T> RootExpectOptions(
-    expectationVerb: Translatable? = null,
-    representationInsteadOfSubject: ((T) -> Any)? = null,
-    reporter: Reporter? = null
-): RootExpectOptions<T> = RootExpectOptions(
-    expectationVerb,
-    representationInsteadOfSubject,
-    reporter
-)
